communication-kb/requirements.txt
python-docx>=1.1.0
openpyxl>=3.1.2
xlrd>=2.0.1
yfinance>=0.2.40
